Separated out db queries.
