+++ src/lib/db.js
@@ -878,28 +878,6 @@ export function searchStaffQuery(queries, min, max) {
   });
 }
 
-export function searchTagsQuery(queries, min, max) {
-  return new Promise(resolve => {
-    let queriesReturned = 0;
-    const results = {};
-    queries.forEach(query => {
-      database
-        .select('slug')
-        .from('tags')
-        .where('name', 'like', `%${query}%`)
-        .limit(max - min + 1)
-        .offset(min)
-        .then(rows => {
-          queriesReturned += 1;
-          results[query] = _.map(rows, row => row.slug);
-          if (queriesReturned >= queries.length) {
-            resolve(results);
-          }
-        });
-    });
-  });
-}
-
 export function searchPostsQuery(queries, min, max) {
   return new Promise(resolve => {
     let queriesReturned = 0;
@@ -1143,149 +1121,6 @@ export function createStaff(staffObject) {
         // Insert statements should throw errors themselves
         // if they fail
         resolve(true);
-      });
-  });
-}
-
-/**
- * Adds a tag to the database.
- * @param {slug, name} tagObject the tag object to insert.
- */
-
-export function addTag(tagObject) {
-  return new Promise(resolve => {
-    database('tags')
-      .insert(tagObject)
-      .then(() => {
-        // Insert statements should throw errors themselves if they fail.
-        resolve(true);
-      });
-  });
-}
-
-/**
- * Queries the tags table in the database for all tags with the given slugs selecting for
- * the given columns.
- * @param  {string[]}    slugs   the tags to select
- * @param  {string[]}    columns the columns of tags to return
- * @returns{Object[]}    rows    the result.
- */
-export function tagQuery(slugs, columns) {
-  if (!_.isArray(slugs) || !_.isArray(columns)) {
-    throw new Error(
-      `tagQuery must be called with slugs and columns as arrays. Received slugs ${slugs} and columns ${columns}`,
-    );
-  }
-
-  return new Promise(resolve => {
-    // Slug is required for the falcor router.
-    const processedColumns = _.uniq(columns.concat(['slug']));
-    database
-      .select(...processedColumns)
-      .from('tags')
-      .whereIn('slug', slugs)
-      .then(rows => {
-        resolve(rows);
-      })
-      .catch(e => {
-        throw new Error(e);
-      });
-  });
-}
-
-/**
- * Update tags given a json graph.
- * @param   {Object}   jsonGraphArg - the json graph to update
- * @returns {boolean}
- */
-export function updateTags(jsonGraphArg) {
-  return new Promise(resolve => {
-    const updatesCalled = Object.keys(jsonGraphArg).length;
-    let updatesReturned = 0;
-    _.forEach(jsonGraphArg, (tagObject, slug) => {
-      database('tags')
-        .where('slug', '=', slug)
-        .update(tagObject)
-        .then(data => {
-          if (data !== 1) {
-            throw new Error(`Problems updating tag with slug: ${slug}`);
-          }
-          updatesReturned += 1;
-          if (updatesReturned >= updatesCalled) {
-            resolve(true);
-          }
-        })
-        .catch(e => {
-          throw e;
-        });
-    });
-  });
-}
-
-/**
- * Updates the tags for a given article.
- * @param   {int}      articleId   The id of the article.
- * @param   {int[]}    newTags     The ids of the tags.
- * @returns {string[]} slugs       The edited slugs.
- */
-export function updateArticleTags(articleId, newTags) {
-  return new Promise(resolve => {
-    database('articles_tags')
-      .where('article_id', '=', articleId)
-      .del()
-      .then(() => {
-        const insertArray = _.map(newTags, tagId => ({
-          article_id: articleId,
-          tag_id: tagId,
-        }));
-        return database('articles_tags').insert(insertArray);
-      })
-      .then(() => {
-        database
-          .select('slug')
-          .from('articles_tags')
-          .innerJoin('tags', 'tags.id', '=', 'articles_tags.tag_id')
-          .where('article_id', '=', articleId)
-          .then(rows => resolve(rows.map(row => row.slug)));
-      });
-  });
-}
-
-/**
- * @param   {string[]} slugs the slugs whose tags we receive
- * @returns {Object}   data  the data
- */
-export function articleTagQuery(slugs) {
-  // slugs function parameter is an array of article slugs
-  // of which to fetch the staff member of.
-  // The function returns an object with article slugs
-  // as keys and values being arrays of author slugs.
-  return new Promise(resolve => {
-    database
-      .select('articles.slug as articleSlug', 'tags.slug as tagSlug')
-      .from('tags')
-      .innerJoin('articles_tags', 'tags.id', '=', 'tag_id')
-      .innerJoin('articles', 'articles.id', '=', 'article_id')
-      .whereIn('articles.slug', slugs)
-      .orderBy('articles_tags.id', 'asc')
-      .then(rows => {
-        // rows is an array of objects with keys authorSlug and articleSlug
-        const data = {};
-        rows.forEach(row => {
-          // This will input them in ascending order by id (which represents time they were
-          // inserted as author of that article) as the query was structured so.
-          if (!has.call(data, row.articleSlug)) {
-            data[row.articleSlug] = [row.tagSlug];
-          } else {
-            data[row.articleSlug].push(row.tagSlug);
-          }
-        });
-        // database.destroy();
-        resolve(data);
-      })
-      .catch(e => {
-        // database.destroy();
-        throw new Error(e);
       });
   });
 }
